Created translated string resource file for Portuguese
A new string resource file with Portuguese text has been created. The strings that do not require translation have been marked as translatable="false" in the original file.

The default language remains English, which is in the original string resource file. However, if a user accesses the Stylo app on a device configured with Brazilian Portuguese, the text in the app will appear translated into the user's language.